Refactor tests to do the slow XML parsing a minimum number of times.
XXX This could probably be simplfied further still, but the current
approach was the least invasive.

Many of the tests used setUp() to parse an XML file and create some
i18n data structure.  The tests only read the data structure, so
there's no need to re-parse the data each time.

The changes reduces the time it takes to run the zope.i18n tests from
several hundred seconds to 17 seconds on my machine.