Set minimum coverage at 99 percent.
We had 100% test coverage previously, but this dropped to around 99.11 percent when using the new default in tox.ini:

```
[coverage:run]
branch = True
```

With branch=False, the coverage would be back at 100 percent,
but I don't see a way to configure this in meta.toml, and it may be better to see the actual coverage.

We should add more tests to cover the untested branches.
For now, we lower our expectations.